# Conflicts:
#	.gitignore
#	pom.xml
#	src/main/java/net/blerf/ftl/constants/FTLConstants.java
#	src/main/java/net/blerf/ftl/model/ShipLayout.java
#	src/main/java/net/blerf/ftl/model/Stats.java
#	src/main/java/net/blerf/ftl/model/sectortree/SectorTree.java
#	src/main/java/net/blerf/ftl/model/sectortree/SectorTreeListener.java
#	src/main/java/net/blerf/ftl/parser/DatParser.java
#	src/main/java/net/blerf/ftl/parser/SavedGameParser.java
#	src/main/java/net/blerf/ftl/parser/random/RandRNG.java
#	src/main/java/net/blerf/ftl/parser/sectormap/RandomSectorMapGenerator.java
#	src/main/java/net/blerf/ftl/parser/sectortree/RandomSectorTreeGenerator.java
#	src/main/java/net/blerf/ftl/ui/ExtensionFileFilter.java
#	src/main/java/net/blerf/ftl/ui/FTLFrame.java
#	src/main/java/net/blerf/ftl/ui/FieldEditorPanel.java
#	src/main/java/net/blerf/ftl/ui/ProfileShipUnlockPanel.java
#	src/main/java/net/blerf/ftl/ui/ReferenceSprite.java
#	src/main/java/net/blerf/ftl/ui/SavedGameFloorplanPanel.java
#	src/main/java/net/blerf/ftl/ui/SavedGameGeneralPanel.java
#	src/main/java/net/blerf/ftl/ui/SavedGameHangarPanel.java
#	src/main/java/net/blerf/ftl/ui/SavedGameSectorMapPanel.java
#	src/main/java/net/blerf/ftl/ui/SavedGameSectorTreePanel.java
#	src/main/java/net/blerf/ftl/ui/SectorMapLayout.java
#	src/main/java/net/blerf/ftl/ui/Statusbar.java
#	src/main/java/net/blerf/ftl/ui/StoreShelfPanel.java
#	src/main/java/net/blerf/ftl/ui/hud/SpriteSelector.java
#	src/main/java/net/vhati/ftldat/FTLPack.java
#	src/main/java/net/vhati/ftldat/FolderPack.java
#	src/main/java/net/vhati/ftldat/PkgPack.java
#	src/main/java/net/vhati/modmanager/core/SloppyXMLParser.java

+++ src/main/java/net/blerf/ftl/ui/SavedGameFloorplanPanel.java
@@ -797,7 +797,7 @@ public class SavedGameFloorplanPanel extends JPanel {
 		int actualDroneSlots;
 
 		if ( shipBlueprint.getDroneSlots() != null ) {
-			int minDroneSlots = shipBlueprint.getDroneSlots().intValue();
+			int minDroneSlots = shipBlueprint.getDroneSlots();
 			if ( droneList.size() > minDroneSlots ) {
 				log.warn( String.format( "Ship state has %d drones, exceeding %d droneSlots on its blueprint", droneList.size(), minDroneSlots ) );
 			}
@@ -936,7 +936,7 @@ public class SavedGameFloorplanPanel extends JPanel {
 					break;
 				}
 			}
-			addBreachSprite( shipBundle, breachX, breachY, roomId, squareId, breachEntry.getValue().intValue() );
+			addBreachSprite( shipBundle, breachX, breachY, roomId, squareId, breachEntry.getValue());
 		}
 
 		// Add fires.
@@ -2925,7 +2925,7 @@ public class SavedGameFloorplanPanel extends JPanel {
 		int maxSystemCapacity = systemBlueprint.getMaxPower();
 		Integer maxPowerOverride = shipBundle.getShipBlueprint().getSystemList().getSystemRoom( systemRef.get().getSystemType() )[0].getMaxPower();
 		if ( maxPowerOverride != null ) {
-			maxSystemCapacity = maxPowerOverride.intValue();
+			maxSystemCapacity = maxPowerOverride;
 		}
 
 		final int batteryCapacity = shipBundle.getBatteryPoolCapacity();
